move DynamicGraph::execute to DynamicPipeline::push
