+++ src/plugins/intel_npu/src/backend/include/zero_dynamic_pipeline.hpp
@@ -98,6 +98,13 @@ public:
                                 const std::shared_ptr<ov::ITensor>& userTensor = nullptr) override;
 
 private:
+    void execute_vm_runtime(npu_vm_runtime_handle_t vmRuntime,
+                            IDynamicGraph::GraphArguments& args,
+                            std::vector<ze_command_list_handle_t>& commandLists,
+                            ze_command_queue_handle_t commandQueue,
+                            ze_fence_handle_t fence,
+                            ze_event_handle_t event);
+
     std::vector<std::unique_ptr<PipelinedCommandLists>> _command_lists;
 };
