+++ src/plugins/intel_npu/src/compiler_adapter/CMakeLists.txt
@@ -8,7 +8,7 @@ file(GLOB_RECURSE SOURCES *.cpp *.hpp *.h)
 
 source_group(TREE ${CMAKE_CURRENT_SOURCE_DIR} FILES ${SOURCES})
 
-add_library(${TARGET_NAME} STATIC ${SOURCES} $<TARGET_OBJECTS:openvino_npu_vcl_utils> $<TARGET_OBJECTS:openvino_npu_vm_utils>)
+add_library(${TARGET_NAME} STATIC ${SOURCES} $<TARGET_OBJECTS:openvino_npu_vcl_utils>)
 
 ov_build_target_faster(${TARGET_NAME}
     PCH_HEADER "src/precomp.hpp"
@@ -28,6 +28,7 @@ target_link_libraries(${TARGET_NAME}
         openvino::npu_al
         openvino::xml_util
         openvino::npu_vm_runtime_api
+        openvino::npu_vm_utils
 )
 
 #
