+++ src/plugins/intel_npu/src/compiler_adapter/src/dynamic_graph.cpp
@@ -34,13 +34,9 @@ public:
     uint64_t getNumSubgraphs() override {
         return _engineProperties.numOfSubGraphs;
     }
-    void executeGraph(const std::shared_ptr<ZeroInitStructsHolder>& zeroInitStruct,
-                      DynamicGraph::GraphArguments& args,
-                      std::vector<ze_command_list_handle_t>& commandLists,
-                      ze_command_queue_handle_t commandQueue,
-                      ze_fence_handle_t inferenceFence,
-                      ze_event_handle_t event,
-                      ze_graph_profiling_pool_handle_t profiling) override;
+    npu_vm_runtime_handle_t getVmRuntimeHandle() const override {
+        return _engine;
+    }
     void getBinding(DynamicGraph::GraphArguments& binding) override;
 
     virtual ~DynamicGraphImpl() {
@@ -297,101 +293,6 @@ void DynamicGraphImpl::setArgumentValueWithStrides(uint32_t argi,
     }
 }
 
-void DynamicGraphImpl::executeGraph(const std::shared_ptr<ZeroInitStructsHolder>& zeroInitStruct,
-                                    IDynamicGraph::GraphArguments& args,
-                                    std::vector<ze_command_list_handle_t>& commandLists,
-                                    ze_command_queue_handle_t commandQueue,
-                                    ze_fence_handle_t fence,
-                                    ze_event_handle_t event,
-                                    ze_graph_profiling_pool_handle_t profiling) {
-    _logger.debug("Start to execute graph with runtime engine");
-    std::shared_ptr<DynamicGraph::GraphArgumentsImpl> argsImpl =
-        args._impl ? std::static_pointer_cast<DynamicGraph::GraphArgumentsImpl>(args._impl)
-                   : std::make_shared<DynamicGraph::GraphArgumentsImpl>();
-
-    bool noTensorChange = true;
-    npu_vm_runtime_execute_params_t* params = &argsImpl->_executeParams;
-    for (auto& in : args._inputs) {
-        std::shared_ptr<DynamicGraph::MemRefTypeImpl> inImpl =
-            std::static_pointer_cast<DynamicGraph::MemRefTypeImpl>(in._impl);
-        if (inImpl == nullptr) {
-            inImpl = std::make_shared<DynamicGraph::MemRefTypeImpl>();
-            in._impl = inImpl;
-        }
-        inImpl->UpdateMemRefHandleStatus(in);
-        if (args._impl == nullptr) {
-            argsImpl->_inputMemRefs.push_back(inImpl->_memRef);
-        } else if (inImpl->_ptrUpdated || inImpl->_shapeUpdated || inImpl->_strideUpdated) {
-            noTensorChange = false;
-        }
-    }
-    for (auto& out : args._outputs) {
-        std::shared_ptr<DynamicGraph::MemRefTypeImpl> outImpl =
-            std::static_pointer_cast<DynamicGraph::MemRefTypeImpl>(out._impl);
-        if (outImpl == nullptr) {
-            outImpl = std::make_shared<DynamicGraph::MemRefTypeImpl>();
-            out._impl = outImpl;
-        }
-        outImpl->UpdateMemRefHandleStatus(out);
-        if (args._impl == nullptr) {
-            argsImpl->_outputMemRefs.push_back(outImpl->_memRef);
-        } else if (outImpl->_ptrUpdated || outImpl->_shapeUpdated || outImpl->_strideUpdated) {
-            noTensorChange = false;
-        }
-    }
-
-    if (args._impl == nullptr || !noTensorChange) {
-        _logger.debug("Reset command list to run with runtime");
-        // Reset commandLists since there are tensor with new shapes or it is the first execution, can not reuse command
-        // list with update
-        for (auto& cmdList : commandLists) {
-            zeCommandListReset(cmdList);
-        }
-    } else {
-        _logger.debug("Reuse command list without update since no tensor change detected");
-
-        auto result = zeCommandQueueExecuteCommandLists(commandQueue,
-                                                        static_cast<uint32_t>(commandLists.size()),
-                                                        commandLists.data(),
-                                                        fence);
-        if (result != ZE_RESULT_SUCCESS) {
-            OPENVINO_THROW("Failed to submit command lists");
-        }
-        return;
-    }
-
-    // Prepare execution context for each graph arguments
-    if (params->executionContext == nullptr) {
-        if (npuVMRuntimeCreateExecutionContext(_engine, &params->executionContext) != NPU_VM_RUNTIME_RESULT_SUCCESS) {
-            OPENVINO_THROW("Failed to create a VM execution context");
-        } else {
-            _logger.debug("Execution context is created successfully.");
-        }
-    }
-
-    params->pInputs = argsImpl->_inputMemRefs.data();
-    params->numOfInputs = static_cast<uint32_t>(argsImpl->_inputMemRefs.size());
-    params->pOutputs = argsImpl->_outputMemRefs.data();
-    params->numOfOutputs = static_cast<uint32_t>(argsImpl->_outputMemRefs.size());
-    params->ctx = zeroInitStruct->getContext();
-    params->device = zeroInitStruct->getDevice();
-    params->graphDdiTableExt = zeroInitStruct->getGraphDdiTable().getImpl();
-    params->commandLists = commandLists.data();
-    params->numCommandLists = static_cast<uint64_t>(commandLists.size());
-    params->commandQueue = commandQueue;
-    params->inferenceFence = fence;
-    params->event = event;
-
-    _logger.debug("Execute graph with runtime engine");
-    if (npuVMRuntimeExecute(_engine, params) != NPU_VM_RUNTIME_RESULT_SUCCESS) {
-        OPENVINO_THROW("Failed to execute VM runtime engine");
-    }
-
-    if (args._impl == nullptr) {
-        args._impl = argsImpl;
-    }
-}
-
 void DynamicGraphImpl::predictOutputShape(std::vector<MemRefType>& inputDescriptors,
                                           std::vector<MemRefType>& outputDescriptors) {
     std::vector<npu_vm_runtime_mem_ref_handle_t> inputs;
@@ -402,7 +303,7 @@ void DynamicGraphImpl::predictOutputShape(std::vector<MemRefType>& inputDescript
             inImpl = std::make_shared<DynamicGraph::MemRefTypeImpl>();
             in._impl = inImpl;
         }
-        inImpl->UpdateMemRefHandleStatus(in);
+        inImpl->updateMemRefHandleStatus(in);
         inputs.push_back(inImpl->_memRef);
     }
     std::vector<npu_vm_runtime_mem_ref_handle_t> outputs;
@@ -413,7 +314,7 @@ void DynamicGraphImpl::predictOutputShape(std::vector<MemRefType>& inputDescript
             outImpl = std::make_shared<DynamicGraph::MemRefTypeImpl>();
             out._impl = outImpl;
         }
-        outImpl->UpdateMemRefHandleStatus(out);
+        outImpl->updateMemRefHandleStatus(out);
         outputs.push_back(outImpl->_memRef);
     }
 
@@ -711,21 +612,6 @@ DynamicGraph::~DynamicGraph() {
     }
 }
 
-void DynamicGraph::execute(const std::shared_ptr<ZeroInitStructsHolder>& zeroInitStruct,
-                           DynamicGraph::GraphArguments& args,
-                           std::vector<ze_command_list_handle_t>& commandLists,
-                           ze_command_queue_handle_t commandQueue,
-                           ze_fence_handle_t inferenceFence,
-                           ze_event_handle_t event,
-                           ze_graph_profiling_pool_handle_t profiling) {
-    auto impl = reinterpret_cast<DynamicGraphImpl*>(_impl.get());
-
-    if (impl == nullptr)
-        return;
-
-    impl->executeGraph(zeroInitStruct, args, commandLists, commandQueue, inferenceFence, event, profiling);
-}
-
 void DynamicGraph::getBinding(GraphArguments& args) {
     auto impl = reinterpret_cast<DynamicGraphImpl*>(_impl.get());
 
@@ -733,6 +619,15 @@ void DynamicGraph::getBinding(GraphArguments& args) {
         return;
 
     impl->getBinding(args);
+}
+
+npu_vm_runtime_handle_t DynamicGraph::get_vm_runtime_handle() const {
+    auto impl = reinterpret_cast<DynamicGraphImpl*>(_impl.get());
+
+    if (impl == nullptr)
+        return nullptr;
+
+    return impl->getVmRuntimeHandle();
 }
 
 uint64_t DynamicGraph::get_num_subgraphs() const {
